Add budgeting, remove old VCs
