+++ Bulldog Bucks/Resources/Cell Models/BalanceCollectionViewCell.swift
@@ -1,0 +1,19 @@
+//
+//  BalanceCollectionViewCell.swift
+//  Bulldog Bucks
+//
+//  Created by Rudy Bermudez on 9/4/17.
+//
+//
+
+import UIKit
+
+class BalanceCollectionViewCell: UICollectionViewCell {
+    
+    public static let reuseIdentifier = "BalanceCollectionViewCell"
+    
+    @IBOutlet weak var amountLabel: UILabel!
+    
+    @IBOutlet weak var weeklyLabel: UILabel!
+    
+}
